Implement cached height config map
